PIM-3595: move referenced collection in MongoDBODM folder
