+++ src/Akeneo/Bundle/StorageUtilsBundle/Resources/config/factories.yml
@@ -1,5 +1,5 @@
 parameters:
-    akeneo_storage_utils.factory.referenced_collection.class: Akeneo\Bundle\StorageUtilsBundle\Doctrine\ReferencedCollectionFactory
+    akeneo_storage_utils.factory.referenced_collection.class: Akeneo\Bundle\StorageUtilsBundle\Doctrine\MongoDBODM\Collections\ReferencedCollectionFactory
 
 services:
     akeneo_storage_utils.factory.referenced_collection:
